Add a mechanism to put information on the tilemap.
Tilemap now holds infomation such as whether it allows item to drop,
permission to put furniture, and etc. When the unity is on editor mode
or runtime, TilemapGridProperties will run and read all the information stored
on tile and store it as an scriptable objects. We can now check the x and y axis of
a specific cell on the map and obtain the necessary permissions and decide what to do.